Split League Classes into seperate modules, Continue tidying and fixing code.
