+++ pyproject.toml
@@ -6,7 +6,7 @@ requires-python = ">=3.9"
 [tool.black]
 line-length = 120
 [tool.ruff]
-lint.select = [
+select = [
     "A",  # prevent using keywords that clobber python builtins
     "B",  # bugbear: security warnings
     "C90", # mccabe: Code Complexity
@@ -19,14 +19,14 @@ lint.select = [
     "RUF",  # the ruff developer's own rules
 ]
 
-lint.ignore = ["D213","D203","PD901","PD011","PTH123"]
+ignore = ["D213","D203","PD901","PD011","PTH123"]
 
 # Allow autofix for all enabled rules (when `--fix`) is provided.
-lint.fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
-lint.unfixable = []
+fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
+unfixable = []
 
 # Exclude a variety of commonly ignored directories.
-lint.exclude = [
+exclude = [
     ".bzr",
     ".direnv",
     ".eggs",
@@ -49,17 +49,17 @@ lint.exclude = [
     "node_modules",
     "venv",
 ]
-lint.per-file-ignores = {"__init__.py" = ["D"]}
+per-file-ignores = {"__init__.py" = ["D"]}
 
 # Same as Black.
 line-length = 120
 
 # Allow unused variables when underscore-prefixed.
-lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
+dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
 
 # Assume Python 3.10.
 target-version = "py310"
 
-[tool.ruff.lint.mccabe]
+[tool.ruff.mccabe]
 # Unlike Flake8, default to a complexity level of 10.
 max-complexity = 10
